VP-1310: Move logic from AuthorizeNetController
- Refactored `AuthorizeNetController` to use `IAuthorizeNetRegisterPaymentService` logic;
