+++ src/VirtoCommerce.AuthorizeNet.Web/Services/IAuthorizeNetRegisterPaymentService.cs
@@ -1,0 +1,10 @@
+using System.Collections.Specialized;
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.AuthorizeNet.Web.Services
+{
+    public interface IAuthorizeNetRegisterPaymentService
+    {
+        Task<string> RegisterPaymentAsync(string orderId, NameValueCollection paymentParameters);
+    }
+}
